+++ src/plugin-window/runtime.js
@@ -529,7 +529,7 @@
 
   function baseHost() {
     return {
-      appVersion:'3.61.49',
+      appVersion:'3.61.50',
       platform:window.DKDSPlatform,
       isAuxiliaryWindow:true,
       isWebClient:false,